# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * foldr.storage API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface User {
  id: string;
  email: string;
  displayName?: string | null;
  walletAddress?: string | null;
  timezone?: string | null;
  totpEnabled?: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export type FileStorageBackend =
  (typeof FileStorageBackend)[keyof typeof FileStorageBackend];

export const FileStorageBackend = {
  lighthouse: "lighthouse",
  fallback: "fallback",
} as const;

export type FileAccessCondition = { [key: string]: unknown } | null;

export interface File {
  id: string;
  userId: string;
  name: string;
  size: number;
  mimeType: string;
  cid: string;
  lighthouseFileId?: string | null;
  storageBackend: FileStorageBackend;
  isEncrypted: boolean;
  isDeleted: boolean;
  accessCondition?: FileAccessCondition;
  createdAt: string;
  updatedAt: string;
}

export interface FileListResponse {
  files: File[];
  total: number;
  page: number;
  limit: number;
}

export interface ShareLink {
  id: string;
  fileId: string;
  token: string;
  url: string;
  expiresAt?: string | null;
  createdAt: string;
}

export interface SharedFileResponse {
  file: File;
  shareLink: ShareLink;
  downloadUrl: string;
}

export interface ShareEncryptedRequest {
  /** Ethereum wallet address of the recipient */
  recipientAddress: string;
}

export interface RevokeAccessRequest {
  /** Ethereum wallet address to revoke access from */
  revokeAddress: string;
}

export type TokenGateConditionReturnValueTest = {
  comparator: string;
  value: string;
};

export interface TokenGateCondition {
  id?: number;
  /** e.g. Ethereum, Polygon, Optimism */
  chain: string;
  /** e.g. balanceOf */
  method: string;
  /** e.g. ERC721, ERC20 */
  standardContractType: string;
  contractAddress: string;
  returnValueTest: TokenGateConditionReturnValueTest;
  parameters: string[];
}

export interface TokenGateRequest {
  conditions: TokenGateCondition[];
  /** Logical grouping of condition ids, e.g. ([1]) */
  aggregator?: string;
}

export interface UpdateProfileRequest {
  displayName?: string | null;
  timezone?: string | null;
}

export interface DeleteAccountRequest {
  password: string;
}

export type StorageStatusBackend =
  (typeof StorageStatusBackend)[keyof typeof StorageStatusBackend];

export const StorageStatusBackend = {
  lighthouse: "lighthouse",
  fallback: "fallback",
} as const;

export interface StorageStatus {
  backend: StorageStatusBackend;
}

export type ListFilesParams = {
  page?: number;
  limit?: number;
  search?: string;
};

export type UploadFileBody = {
  file: Blob;
  encrypt?: boolean;
};

export type CreateShareLinkBody = {
  /** Hours until link expires (omit for no expiry) */
  expiresInHours?: number;
};