# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: foldr.storage API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/profile:
    put:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile (display name, timezone)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/account:
    delete:
      operationId: deleteAccount
      tags: [auth]
      summary: Permanently delete the current user account
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAccountRequest"
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Not authenticated or wrong password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/revoke-sessions:
    post:
      operationId: revokeSessions
      tags: [auth]
      summary: Sign out of all sessions (invalidates all active tokens)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: All sessions revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /files/{id}/decrypt:
    get:
      operationId: decryptFile
      tags: [files]
      summary: Decrypt and stream an encrypted file
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Decrypted file stream
        "400":
          description: File is not encrypted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /share/{token}:
    get:
      operationId: getSharedFile
      tags: [share]
      summary: Access a publicly shared file (no auth required)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shared file info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedFileResponse"
        "404":
          description: Share link not found or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: string
          nullable: true
        walletAddress:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
        totpEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - createdAt
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message
    File:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        size:
          type: integer
        mimeType:
          type: string
        cid:
          type: string
        lighthouseFileId:
          type: string
          nullable: true
        storageBackend:
          type: string
          enum: [lighthouse, fallback]
        isEncrypted:
          type: boolean
        isDeleted:
          type: boolean
        accessCondition:
          type: object
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - size
        - mimeType
        - cid
        - storageBackend
        - isEncrypted
        - isDeleted
        - createdAt
        - updatedAt
...
    ShareLink:
      type: object
      properties:
        id:
          type: string
        fileId:
          type: string
        token:
          type: string
        url:
          type: string
        expiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fileId
        - token
        - url
        - createdAt
    SharedFileResponse:
      type: object
      properties:
        file:
          $ref: "#/components/schemas/File"
        shareLink:
          $ref: "#/components/schemas/ShareLink"
        downloadUrl:
          type: string
      required:
        - file
        - shareLink
        - downloadUrl
...
    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
    DeleteAccountRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
...