# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * foldr.storage API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  email: string;
  displayName?: string | null;
  walletAddress?: string | null;
  timezone?: string | null;
  totpEnabled?: boolean;
  createdAt: Date;
}